Update Chatting Participant
- 채팅방 참가자 카운트 수정

+++ src/main/java/com/smingsming/chat/domain/chat/vo/RoomVo.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class RoomVo {
     private String roomId;
     private String roomName;
-    private int participant;
+    private Long participant;
     private boolean isLock;
     private String userId;
     private String userName;
